Introduce a "protocol" for CrosEC
This protocol is associated with an interface that has two members,
"SendCommand" and "ReadMemory". All of the existing in-tree drivers have
been ported to (1) register this protocol and (2) register a device path
protocol instance that allows for collision detection.

CrosECLib has been updated to detect the presence of a Microchip EC
before completing registration. The Null implementation of CrosECLib
will *always* register, unless another EC comms library has registered.

It is built this way so that:

- A platform can customize on a per-application basis which EC
  implementation is used
- Individual applications can ship with support for multiple ECs and the
  ability to autodetect which one is present.
- Somebody could eventually ship a driver, CrosECDriverDxe (perhaps),
  that provides the same protocol at the platform level and is
  automatically started up and made available for future applications.