ECTool: check that the current board matches the image before flashing
We'll try to read the version from the EC and use that to determine what
the current board is -- typically, the version number is of the format
board_x.y.z; if we run up to the first underscore and truncate the
version string there, we stand a good chance of guessing the board.

We can do the same thing for the firmware image we just read off disk.
For the Microchip MEC images the Framework Laptop uses, the version
number is always going to be in the same place: right after the reset
vectors. I hardcoded an offset (which I somewhat do not like to do) and
we'll read the version from there.

If you really know what you're doing you can specify -f twice to
override the board check. We'll still query the version, but we won't do
anything with what we find.