ECTool: Use the actual flash map to find the firmware ID
We used to hardcode the offset to the version number in the MCHP
firmware images. It turns out that EC images come with a flash section
map that contains all of the requisite version numbers as separate
sections.

In this case, we'll look for RO_FRID. There's also RW_FWID, but it's
redundant for our purposes.