ECTool: use built-in flash maps to avoid VPD and SPI Flags
Instead of hardcoding the RO and RW sizes and offsets for the hx20/hx30
boards, instead look them up in a table based on board name.

This allows us to be aware of the SPI flags region for Azalea and the
VPD regions for the hx20/hx30.

Regions are accumulated in two bit fields:
- One indicating which regions are present
- One indicating which regions are *desired*

The set of desired regions starts out as "all regions", and regions
marked with the SKIP flag are unmarked as desired.

Passing --rw or --ro resets the desire mask to 0 and searches the region
table for RW and RO respectively to mark them as desired.

If we can't figure out what board we're working with, we'll bail out.
However, if the user passes `-f -f` we will force it through with a
50/50 flash layout for a 512kb chip.

We take that bitfield and enumerate it from the highest bit to the
lowest; so doing ensures that we write the RW region *first* in most
cases, and if it fails we can bail out early and not harm the RO region.