CrosECDriver: add a driver that enumerates all compiled-in EC support
The main differences between having ECTool/ECTest do this and having a
driver do are that (1) a platform builder can include a CrosECDriver
linked with whichever EC support library they want, and (2) it will stay
loaded for the duration of system execution.

2 is particularly important for CrosECNullLib, which has "durable" flash
(that is: a memory allocation) that it will reuse.

Use it with:

LOAD FS0:\CrosECDriver.efi